Refactoring to eliminate redundancy/improve organization
