+++ server/__init__.py
@@ -1,5 +1,4 @@
-from flask import Flask, session, jsonify, Response, request, json, render_template, redirect, current_app
-from werkzeug.security import generate_password_hash, check_password_hash
+from flask import Flask
 import os
 import redis
 
@@ -8,8 +7,12 @@ app = Flask(__name__)
 # redis
 db = redis.StrictRedis(host='localhost', port=6379, db=0)
 app.secret_key = os.urandom(24)
-import server.api
 
+import server.registrar
+import server.transit
+import server.dining
+import server.buildings
+import server.directory
 
 if __name__ == '__main__':
     app.run(debug=True)
